Added language probing. Added REPLACE value to Position enum. Reports go at 12:00. Added message deletion. Added graceful shutdown handler. Fixed TG compatability issues. Added GCP required code. Moved properties outside the jar.

+++ src/main/java/net/aregism/trdelnikpolice/model/dto/TelegramMessageDto.java
@@ -8,7 +8,7 @@ import lombok.*;
 @AllArgsConstructor
 public class TelegramMessageDto {
     private long chatId;
-    private long messageId;
+    private int messageId;
     private String text;
     private String sender;
     private boolean replyToBot;
